% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_seminribm.R
\name{seminribm_run}
\alias{seminribm_run}
\title{Run one simulation of a stochastic SEmInR model}
\usage{
seminribm_run(horizon, popSize, R0, latent_mean, infectious_mean, nE, nI,
  initInfectious, doExact, timeStepTauLeap = 0.1, rnd_seed = 1234,
  calc_WIW_Re = FALSE)
}
\arguments{
\item{horizon}{Numeric. Time horizon of the simulation.}

\item{popSize}{Integer. Population size.}

\item{R0}{Numeric. Basic reproduction number.}

\item{latent_mean}{Numeric. Mean latent (infected but not infectious) duration.}

\item{infectious_mean}{Numeric. Mean infectious duration.}

\item{nE}{Integer. Number of 'E' compartments.}

\item{nI}{Integer. Number of 'I' compartments.}

\item{initInfectious}{Integer. Number of individuals initially infected (at time 0).}

\item{doExact}{Boolean. TRUE: the stochastic simulation follows the standard Gillespie algorithm. FALSE: tau-leap approximation is used.}

\item{timeStepTauLeap}{Numeric. Time step for the tau-leap approximation. Default = 0.1. Only used if \code{doExact=FALSE}.}

\item{rnd_seed}{Integer. Seed for the random number generator. Default = 1234.}

\item{calc_WIW_Re}{Boolean. Calculate the 'Who Infected Who' matrice. Warning: slows down simulation. Default = FALSE.}
}
\value{
A named list storing simulation outputs. 
\itemize{
\item \code{times}: simulated calendar times.
\item \code{S}: Time series of the number of susceptible individuals.
\item \code{prev}: Time series of prevalence (number of infectious individuals).
\item \code{R}: Time series of the number of recovered individuals.
\item \code{acq_times}: Disease acquisition times for all infected individuals.
\item \code{acqTransm_times}: Disease acquisition times for all infected individuals that transmitted the disease to at least one other individual.
\item \code{GI_bck}: Backward generation intervals for all infected individuals.
\item \code{GI_fwd}: Forward generation intervals for all infected individuals. One individual may have more than one forward generation interval, hence this object is a list of vectors.
\item \code{Reff}: Effective disease reproduction. Number of secondary cases generated by each infected individual.
}
}
\description{
Stochastic, individual-based, SEIR model with respectively m and n 'E' and 'I' chained compartments. Use Gillespie algorithm, with or without the tau-leap approximation.
}
